Take folders into account when shelving/unshelving

+++ src/main/java/org/jvnet/hudson/plugins/shelveproject/UnshelveProjectExecutable.java
@@ -3,14 +3,18 @@ package org.jvnet.hudson.plugins.shelveproject;
 import hudson.FilePath;
 import hudson.model.Queue;
 import jenkins.model.Jenkins;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.io.FilenameUtils;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
-import java.util.Collection;
+import java.nio.file.Path;
+import java.util.Properties;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import static org.jvnet.hudson.plugins.shelveproject.ShelveProjectExecutable.*;
+import static org.jvnet.hudson.plugins.shelveproject.ShelveProjectExecutable.PROJECT_PATH_PROPERTY;
 
 public class UnshelveProjectExecutable
     implements Queue.Executable
@@ -34,12 +38,24 @@ public class UnshelveProjectExecutable
 
     public void run() {
         for (String shelvedProjectArchiveName : shelvedProjectArchiveNames) {
-            final File shelvedProjectArchive = getArchiveFile(shelvedProjectArchiveName);
+            final File shelvedProjectArchive = DeleteProjectExecutable.getArchiveFile(shelvedProjectArchiveName);
             LOGGER.info("Unshelving project [" + shelvedProjectArchiveName + "].");
+            boolean correctlyExploded;
             try {
-                new FilePath(shelvedProjectArchive).unzip(
-                        new FilePath(new File(Jenkins.getInstance().getRootDir(), "jobs")));
-                Files.delete(shelvedProjectArchive.toPath());
+                if(ARCHIVE_FILE_EXTENSION.equals(FilenameUtils.getExtension(shelvedProjectArchiveName))) {
+                    correctlyExploded = explode(shelvedProjectArchive);
+                    if(correctlyExploded) {
+                        Files.delete(ShelvedProject.getMetadataFileFromArchive(shelvedProjectArchive));
+                    }
+                } else {
+                    legacyExplode(shelvedProjectArchive);
+                    correctlyExploded = true;
+                }
+                if(correctlyExploded) {
+                    Files.delete(shelvedProjectArchive.toPath());
+                } else {
+                    LOGGER.log(Level.INFO, "Skipping deletion of the backup at " + shelvedProjectArchiveName);
+                }
                 Jenkins.getInstance().reload();
             } catch (Exception e) {
                 LOGGER.log(Level.SEVERE, "Could not unarchive project archive [" + shelvedProjectArchiveName + "].", e);
@@ -47,18 +63,26 @@ public class UnshelveProjectExecutable
         }
     }
 
-    private File getArchiveFile(String shelvedProjectArchiveName) {
-        // JENKINS-8759 - The archive name comes from the html form, so take a bit extra care for security reasons by
-        // only accessing the archive if it exists in the directory of shevled projects.
-        File shelvedProjectsDirectory = new File(Jenkins.getInstance().getRootDir(), ShelvedProjectsAction.SHELVED_PROJECTS_DIRECTORY);
-        Collection<File> files = FileUtils.listFiles(shelvedProjectsDirectory, null, false);
-        for (File file : files) {
-            if (StringUtils.equals(file.getName(), shelvedProjectArchiveName)) {
-                return file;
-            }
+    private boolean explode(File shelvedProjectArchive) throws IOException, InterruptedException {
+        File rootDir = Jenkins.getInstance().getRootDir();
+        Properties metadata = ShelvedProject.loadMetadata(shelvedProjectArchive);
+        String projectPathProperty = metadata.getProperty(PROJECT_PATH_PROPERTY);
+        Path projectPath = rootDir.toPath().resolve(projectPathProperty);
+        if (Files.exists(projectPath)) {
+            LOGGER.log(Level.INFO, "A project exist for the given path " + projectPathProperty + "...skipping");
+            return false;
         }
-        return null; // Project was already unshelved?
+        new FilePath(shelvedProjectArchive).untar(
+                new FilePath(new File(rootDir, "jobs")), FilePath.TarCompression.NONE);
+        return true;
     }
+
+    private void legacyExplode(File shelvedProjectArchive) throws IOException, InterruptedException {
+        new FilePath(shelvedProjectArchive).unzip(
+                new FilePath(new File(Jenkins.getInstance().getRootDir(), "jobs")));
+    }
+
+
 
     public long getEstimatedDuration()
     {
